text2speech/requirements.txt
azure-cognitiveservices-speech>=1.34.0
pygame>=2.0.0
python-dotenv>=1.0.0 